requirements.txt
pyrogram
tgcrypto
pytz
psutil
motor
dnspython
GitPython
pykeyboard
bs4
Pillow==8.3.2
pymediainfo
opencv-python-headless 
pytgcalls==3.0.0.dev22
requests
unidecode
pytimeparse
attrify
pySmartDL
html2text
removebg
aiofiles
aiohttp
gpytranslate
gtts
telegraph
pyromod==1.5
openai==0.28
youtube-search-python
pymongo
yt-dlp
wget
python-dateutil
emoji
heroku3
